* Added makefile for tests * Fixed bug in release regarding boundary checks * Settled for cascade limit ~ 1500000 * Changed reference count and size members in object record to unsigned long * Added overflow control structure, rc_overflow()

+++ src/refmem.h
@@ -2,10 +2,11 @@
 
 #include <stddef.h>
 #include <stdbool.h>
+#include <limits.h>
 
 typedef void *obj;
 typedef void(*function1_t)(obj);
-typedef short rc_format;
+typedef unsigned long rc_format;
 
 void retain(obj);
 void release(obj);
